refactor duplicated code across mini games
