+++ src/main/java/work/lclpnet/ap2/api/game/sink/IntDataSink.java
@@ -1,0 +1,8 @@
+package work.lclpnet.ap2.api.game.sink;
+
+public interface IntDataSink<T> {
+
+    void addScore(T subject, int add);
+
+    int getScore(T subject);
+}
